Add Checkpointing for SsdCache
Adds a mechanism for writing a recoverable checkpoint every so many
bytes written into SSD cache.  The checkpoint records the cache size,
the access frequencies, the ids of files present in the cache and the
file id, offset to SSD cache location map.  When evictions take place
after a checkpoint, the cleared regions are logged synchronously
before allowing overwrite. In this way, if on restart we find a well
formed checkpoint file and an eviction log, we can restore the cache
state . The entries that fall on evicted regions are not included in
the recovery. If logging or writing a checkpoint file fails, the
checkpoint data is deleted and no further checkpointing takes place.

When a cache shard is initialized, the we check for existence of a
checkpoint file and eviction log. If both exist and the log is well
formed, each SsdFile in the cache is initialized from its own
checkpoint. If the file is not well formed the checkpoint data is
deleted and the cache shard starts from an empty state.