Fix Mapped Pages Counting for allocateContiguous()
When allocating pages with MmapAllocator::allocateContiguous, we first
check that numAllocated_ has room for the number of pages that are
needed after releasing collateral and the previous contents of the
allocation.  If so, the allocation is possible but may need pages that are backing freed allocations in the different size classes to be advised away.

This was not taking place in if there were existing contiguous allocations that were counted in numExternalMapped_.

We change the meaning of numMapped_ to be the total number of pages
that are backed by memory, whether in the size classes or outside with
allocateContiguous (anonymous standalone mmap). numExternalMapped_ is
not the number of outstanding pages in standalone mmaps from
allocateContiguous but it is included in the numMapped_ count. In this
way, given capacity_, numMapped_ and the number of pages that were
allocated without backing memory, we always know how much we must
advise away from freed memory backed allocations.  This capacity is
available since the total to be covered is never over capacity_.