Use an interleaved tags-pointers-tags-... layout for hash tables
Organizes join and group by hash tables into chunks of 2 cache lines
with 16 tag bytes followed by 16 48 bit pointers. This is analogous to
F14 layout where groups of 16 8 bit tags alternate with payload.

In principle, for hash tables that miss often and where the miss can
be detected with tags alone, putting the tags all together could be
better for locality. Experimentally, we do not see much difference, in
any case not enough to maintain two hash table layouts. On the other
hand, when interleaving tags and pointers, we quite often save a cache
miss for a hit because the tag is in the same cache line as the
pointer. This is seen to be ~15% faster than a layout with separate
tags and pointers. See hashTableBenchmark.cpp for probe level clock
counts with different table sizes and hit rates.

The hash table arrays are slightly smaller, at 8 bytes per entry
instead of 9 bytes wen the tags were separate.