Add background profiler
Adds a profile option that starts Linux perf on the Velox process.
Flags control the duration of profiling. If --profile_min_cpu_pct is
exceeded for up to profiler__check_interval_seconds, perf is
started. Perf is stopped and results are written to the path given to
startProfiler() after --profiler_max_sample_seconds or after CPU falls
below the minimum for a full check interval. If the CPU was above minimum utilization for less than --profiler_min_sample_seconds, the result is not saved to the result directory.

Files are named based on the hostname, datetime of sample end and pid.