Add runner for local distributed execution
Adds LocalRunner and LocalSchema for running distributed Velox plans
in-process with multiple fragments and multiple tasks per
fragment. This introduces the abstractions for distributed plans,
their options, schema and split sources.

Adds an extension of PlanBuilder for building plan trees with
shuffles. These get partitioned into multiple plan trees for execution
with LocalRunner.

Adds LocalRunnerTestBase, a test base class with utilities for
generating multiple tables each with multiple files of test
data. These are then returned as a LocalSchema that can be used to
produce splits for LocalRunner.

Adds error propagation for TaskCursor since distributed plans may have
to produce errors from any fragment that are consumed by the
application via TaskCursor.

This will be generalized to distributed execution on a cluster with Velox coordination and shuffles.